+++ src/main/java/eu/hinsch/spring/boot/actuator/metric/ExecutionMetricAutoConfiguration.java
@@ -9,9 +9,9 @@ import org.springframework.context.annotation.Configuration;
  * Created by lh on 28/04/15.
  */
 @Configuration
-public class ExecutorMetricAutoConfiguration {
+public class ExecutionMetricAutoConfiguration {
     @Bean
-    public DurationMetricAspect durationMetricAspect(GaugeService gaugeService, CounterService counterService) {
-        return new DurationMetricAspect(gaugeService, counterService);
+    public ExecutionMetricAspect durationMetricAspect(GaugeService gaugeService, CounterService counterService) {
+        return new ExecutionMetricAspect(gaugeService, counterService);
     }
 }