+++ src/test/java/eu/hinsch/spring/boot/actuator/metric/SpringBootExecutionMetricTestApplication.java
@@ -1,8 +1,10 @@
 package eu.hinsch.spring.boot.actuator.metric;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.logging.LogLevel;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -20,12 +22,26 @@ public class SpringBootExecutionMetricTestApplication {
         SpringApplication.run(SpringBootExecutionMetricTestApplication.class, args);
     }
 
+    @Autowired
+    private TestInterface test;
+
     @RequestMapping
     @ResponseBody
-    @DurationMetric(value = "test-metric", loglevel = LogLevel.WARN)
+    @ExecutionMetric(value = "test-metric", loglevel = LogLevel.WARN)
     public String aopTest(@RequestParam("param") String param) {
+        test.someMethod();
         return param + callSomeExternalSystem();
     }
+
+    @Component
+    public static class TestClass implements TestInterface {
+        @Override
+        @ExecutionMetric(value = "some-method", loglevel = LogLevel.WARN)
+        public void someMethod() {
+        }
+    }
+
+    public interface TestInterface { void someMethod(); }
 
     private void simulateActivity() {
         try {